Should only download Cygwin's setup.ini once per run
Downloading should be done in the default recipe and not by apt-cyg
for each dependency (which is what happens now). This must then be
put where apt-cyg looks for it.

It is possible to specify which mirror to use and where to place
the cache through these attributes:

node["apt-cyg"]["mirror"]
node["apt-cyg"]["cache_dir"]

Fixes GH-1